Add files via upload
Every accident on our roads is a preventable tragedy. According to www.brake.org.uk 5 people die everyday in the UK and 84 are seriously injured, with majority being car occupants and pedestirans.
Many would think that majority of accidents are caused by reckless and drunk driving. A recent study however shows that a third of accidents occur as a result of failure to look properly. 
There are numerous causes of accidents, but this dataset examines overlooked contributory factors i.e. weather and lighting. 
Together, let's explore the accident statistics from 2019-2022.